+++ lib_acl_cpp/include/acl_cpp/mime/mime_image.hpp
@@ -10,7 +10,7 @@ class ACL_CPP_API mime_image : public mime_node {
 public:
 	mime_image(const char* emailFile, const MIME_NODE* node,
 		bool enableDecode = true, const char* toCharset = "gb2312",
-		off_t off = 0)
+		long long off = 0)
 		: mime_node(emailFile, node, enableDecode, toCharset, off)
 	{
 	}
